WeldArray creation optimization + output optimization
For weldarray creation, moved all the unrequired dicts/variables to be
global variables instead of per WeldArray variables. (improves overhead
of creating new weldarrays as compared to np arrays)

If in place op, i.e, a += b, then new weld array won't be created.
(Note this does not apply to a = a + b)