Fixes vectorization bug
*Originallly, after vectorization, we were getting code like:

  (let a=(e0);(let a#1=((let a=((let a=(e1);result(
    for(
      fringeiter(a),
      for(
        simditer(a),
        appender[f32],
        |b,i,x|
          merge(b,(sqrt(x)))
      ),
      |b,i,x|
        merge(b,(sqrt(x)))
    )
  )));(let a#1=(e2);result(...
  ...
  ))))))

So variable names were being reused. Re-using the variable names seems
to work in some cases - the scoping rules aren't entirely clear to me...
but by declaring the symbol_generator before the transformation calls,
we are able to get unique symbol names for the let statements (a,
a#1, a#2 etc.) - which fixes the tets "blackscholes_bug", and
"vectorization_bug", from the numpy tests.