Fixed views bug introduced by obj_id change
* 73 / 76 tests passing. Failing tests were supposed to fail for now --
need to fix 1d striding, and non-contig array special cases.
* Originally, used to insert the weld_code directly. This was changed to
the nicer method of adding weldobj.obj_id - and weldobject adding the
latest code in evaluate. This caused views to completely break -- now it
seems to be fixed. A slightly odd consequence is the need to always call
view.evaluate when passing to np.allclose (this was not required for
non-views usually)
* Minor bugfix when views were evaluated and _weldarray_view was not
preserved + new test.
* Minor cleanup across the weldarray class