LLVM Single Threaded Backend (#377)
New single threaded backend based on LLVM APIs.

This commit is *backwards-compatible* with the existing backend and can be enabled or disabled via a Weld flag. However, future commits on this branch are *not* guaranteed to maintain backwards API compatibility with master.

See the README for which features are unsupported in this backend.